created wallet servive and controller

+++ moneyLogBackend/src/main/java/com/mukesh/moneyLogBackend/Repository/UserRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepo extends JpaRepository<User, Integer> {
+public interface UserRepo extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
     Optional<User> findByActivationToken(String activationToken);
     Optional<User> findByUsername(String username);
